+++ src/controls/image_frame.rs
@@ -19,8 +19,11 @@
 */
 use std::hash::Hash;
 use std::any::TypeId;
+use std::marker::PhantomData;
+use std::ptr;
 
-use winapi::{HWND, HANDLE};
+use winapi::{HWND, HANDLE, WPARAM, IMAGE_BITMAP};
+use user32::SendMessageW;
 
 use ui::Ui;
 use controls::{Control, ControlT, ControlType, AnyHandle, HandleSpec};
@@ -52,8 +55,8 @@ pub struct ImageFrameT<ID: Hash+Clone> {
     pub parent: ID
 }
 
-impl<ID: Hash+Clone> ControlT<ID> for ImageFrameT<ID> {
-    fn type_id(&self) -> TypeId { TypeId::of::<ImageFrame>() }
+impl<ID: Hash+Clone+'static> ControlT<ID> for ImageFrameT<ID> {
+    fn type_id(&self) -> TypeId { TypeId::of::<ImageFrame<ID>>() }
 
     fn events(&self) -> Vec<Event> {
         vec![Destroyed, Moved, Resized, Click, DoubleClick, Event::Any]
@@ -63,7 +66,6 @@ impl<ID: Hash+Clone> ControlT<ID> for ImageFrameT<ID> {
         use low::window_helper::{WindowParams, build_window, handle_of_window};
         use low::defs::{SS_NOTIFY, SS_BITMAP};
         use winapi::{DWORD, WS_VISIBLE, WS_DISABLED, WS_CHILD};
-        use std::ptr;
 
         let flags: DWORD = WS_CHILD | SS_NOTIFY |  SS_BITMAP | 
         if self.visible    { WS_VISIBLE }   else { 0 } |
@@ -101,7 +103,7 @@ impl<ID: Hash+Clone> ControlT<ID> for ImageFrameT<ID> {
         match unsafe{ build_window(params) } {
             Ok(h) => {
                 unsafe{ set_image(h, image); }
-                Ok( Box::new(ImageFrame{handle: h}) )
+                Ok( Box::new(ImageFrame::<ID>{handle: h, p: PhantomData}) )
             },
             Err(e) => Err(Error::System(e))
         }
@@ -112,11 +114,12 @@ impl<ID: Hash+Clone> ControlT<ID> for ImageFrameT<ID> {
 /**
     A frame that display a bitmap image loaded from a Image resource.
 */
-pub struct ImageFrame {
-    handle: HWND
+pub struct ImageFrame<ID: Hash+Clone> {
+    handle: HWND,
+    p: PhantomData<ID>
 }
 
-impl ImageFrame {
+impl<ID: Hash+Clone> ImageFrame<ID> {
     pub fn get_visibility(&self) -> bool { unsafe{ ::low::window_helper::get_window_visibility(self.handle) } }
     pub fn set_visibility(&self, visible: bool) { unsafe{ ::low::window_helper::set_window_visibility(self.handle, visible); }}
     pub fn get_position(&self) -> (i32, i32) { unsafe{ ::low::window_helper::get_window_position(self.handle) } }
@@ -125,9 +128,51 @@ impl ImageFrame {
     pub fn set_size(&self, w: u32, h: u32) { unsafe{ ::low::window_helper::set_window_size(self.handle, w, h, false); } }
     pub fn get_enabled(&self) -> bool { unsafe{ ::low::window_helper::get_window_enabled(self.handle) } }
     pub fn set_enabled(&self, e:bool) { unsafe{ ::low::window_helper::set_window_enabled(self.handle, e); } }
+
+    /// Set the image of image frame  
+    /// Pass `None` as an argument to remove the image.
+    pub fn set_image(&self, ui: &Ui<ID>, img: Option<&ID>) -> Result<(), Error> {
+        if !ui.has_handle(&self.handle()){
+            return Err(Error::BadUi("Image resource and control must be in the same Ui.".to_string()));
+        }
+
+        let img_handle = if let Some(id) = img {
+            match ui.handle_of(id) {
+                Ok(h) => match h {
+                    AnyHandle::HANDLE(j, HandleSpec::Bitmap) => j,
+                    h => { return Err(Error::BadResource(format!("An Image resource is required, got {:?}", h))) }
+                },
+                Err(e) => { return Err(e); }
+            }
+        } else {
+            ptr::null_mut()
+        };
+
+        unsafe{
+            set_image(self.handle, img_handle);
+        }
+
+        Ok(())
+    }
+
+    /// Return the resource identifier of the control image    
+    /// Return `None` if the control do not have an image. Will also return `None` if the UI is unrelated to the control.
+    pub fn get_image(&self, ui: &Ui<ID>) -> Option<ID> {
+        use low::defs::STM_GETIMAGE;
+
+        let image_handle = unsafe{ SendMessageW(self.handle, STM_GETIMAGE, IMAGE_BITMAP as WPARAM, 0) as HANDLE };
+        if image_handle.is_null() {
+            return None;
+        }
+
+        match ui.id_from_handle(&AnyHandle::HANDLE(image_handle, HandleSpec::Bitmap)) {
+            Ok(id) => Some(id),
+            Err(_) => None
+        }
+    }
 }
 
-impl Control for ImageFrame {
+impl<ID: Hash+Clone> Control for ImageFrame<ID> {
 
     fn handle(&self) -> AnyHandle {
         AnyHandle::HWND(self.handle)
@@ -148,8 +193,7 @@ impl Control for ImageFrame {
     Set the image of an image frame
 */
 unsafe fn set_image(hwnd: HWND, image: HANDLE) {
-    use user32::SendMessageW;
-    use winapi::{IMAGE_BITMAP, LPARAM, WPARAM};
+    use winapi::LPARAM;
     use low::defs::STM_SETIMAGE;
 
     SendMessageW(hwnd, STM_SETIMAGE, IMAGE_BITMAP as WPARAM, image as LPARAM);