+++ CHANGELOG.md
@@ -6,8 +6,13 @@ For more information also see:
   * https://gabdube.github.io/native-windows-gui/book/events.html (events basics)
   * https://gabdube.github.io/native-windows-gui/book/custom_events.html (defining custom events)
   * https://github.com/gabdube/native-windows-gui/blob/master/examples/templating.rs (simple example)
-* The ability to load bitmap, ico and cursor resources in the **Image resource**
-* The **ImageFrame** control to display a bitmap in a window
+* New resource:
+  * The **Image** resource to load bitmap, ico and cursor files ()
+* New controls:
+  * The **ImageFrame** control to display a bitmap in a window ()
+* New methods for `UI`:
+  * `has_handle`: Check if the ui has an object identified by an handle
+  * `id_from_handle`: Return the `ID` associated with an HANDLE
 
 # BETA 1 (0.2.0)